Data Loader change + debug mode
Changed the majority of edf loading to mne method. It is ~4x faster than pyedflib. Pyedflib is still used to get the header info though, as that is 10x faster than mne methods.

Also added a debug mode to not save any outputs. Allows new code changes to be tested easily.